First stab into the reworked version
Just merged Type 1 fonts with minimal layout feature for now.

Everything is build with AFDKO this time, just an experiment to see how
far I can go with this. The PFA files are not kept under version
control, however, as they are not version control friendly, and instead
disassembled versions by t1disasm are kept.